% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/primitives.R
\name{add_edge}
\alias{add_edge}
\alias{add_edges}
\alias{add_edge.datagraph_edge}
\alias{add_edge.datagraph_vertex}
\alias{add_edge.character}
\alias{add_edge.list}
\alias{add_edges.datagraph}
\alias{add_edges.data.table}
\title{Add edges (between vertices) to a graph}
\usage{
add_edge(x, ...)

add_edges(x, ...)

\method{add_edge}{datagraph_edge}(x, graph)

\method{add_edge}{datagraph_vertex}(from, to, graph, data = NULL)

\method{add_edge}{character}(from, to, graph, data = NULL)

\method{add_edge}{list}(x, graph)

\method{add_edges}{datagraph}(x, edges)

\method{add_edges}{data.table}(x, graph)
}
\arguments{
\item{x}{see S3 methods}

\item{...}{passed to individual methods}

\item{graph}{a datagraph object}

\item{from}{character; id of the start node}

\item{to}{character; id of the end node}

\item{data}{list of data attributes}
}
\value{



}
\description{
Add edges (between vertices) to a graph
}
