% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/primitives.R
\name{add_vertex}
\alias{add_vertex}
\alias{add_vertices}
\alias{add_vertex.datagraph}
\alias{add_vertex.character}
\alias{add_vertex.data.table}
\alias{add_vertex.list}
\alias{add_vertices.datagraph}
\alias{add_vertices.data.table}
\title{Add vertices}
\usage{
add_vertex(x, ...)

add_vertices(x, ...)

\method{add_vertex}{datagraph}(x, vertex)

\method{add_vertex}{character}(x, graph, from = NULL, to = NULL, ...)

\method{add_vertex}{data.table}(x, graph)

\method{add_vertex}{list}(x, graph)

\method{add_vertices}{datagraph}(x, vertices)

\method{add_vertices}{data.table}(x, graph)
}
\arguments{
\item{x}{see the S3 methods}

\item{...}{args passed to the S3 methods}

\item{vertex}{id of a vertex}

\item{graph}{datagraph object}

\item{from}{vector of vertices to connect from}

\item{to}{vector of vertices to connect from}

\item{vertices}{vector of new ids}
}
\description{
Add vertices
}
