% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environments.R
\name{all.equal.datagraph}
\alias{all.equal.datagraph}
\title{#' @export
vertex.default <- function() {
  obj <- new.env()
  class(obj) <- c("datagraph_vertex")
  return(obj)
}}
\usage{
\method{all}{equal.datagraph}(current, target)
}
\description{
#' @export
vertex.character <- function(id, from = NULL, to = NULL, ..., graph = emptyenv()) {
  obj <- list2env(list(id = id, from = from, to = to, data = list(...)), parent = graph)
  class(obj) <- c("datagraph_vertex")
  return(obj)
}
}
