% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{are_adjacent}
\alias{are_adjacent}
\title{Test if two vertices are adjacent (direct neighbors)}
\usage{
are_adjacent(x, ...)
}
\arguments{
\item{x}{see the S3 methods}

\item{...}{args passed to the S3 methods}
}
\description{
Test if two vertices are adjacent (direct neighbors)
}
