% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{as.datagraph}
\alias{as.datagraph}
\title{Constructors of datagraph objects}
\usage{
as.datagraph(x, ...)
}
\arguments{
\item{x}{see the S3 methods}

\item{...}{args passed to the S3 methods}
}
\description{
Constructors of datagraph objects
}
