% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/others.R
\name{check}
\alias{check}
\alias{check.datagraph}
\title{Check/test a datagraph object's consistency}
\usage{
check(x, ...)

\method{check}{datagraph}(x)
}
\arguments{
\item{x}{see the S3 methods}

\item{...}{args passed to the S3 methods}
}
\description{
Check/test a datagraph object's consistency
}
