% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{collapse_vertices}
\alias{collapse_vertices}
\title{Collapse vertices}
\usage{
collapse_vertices(x, ...)
}
\arguments{
\item{x}{see the S3 methods}

\item{...}{args passed to the S3 methods}
}
\description{
Collapse vertices
}
