% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{detect_cycles}
\alias{detect_cycles}
\title{Test that graph contains cycles}
\usage{
detect_cycles(x, ...)
}
\arguments{
\item{x}{see the S3 methods}

\item{...}{args passed to the S3 methods}
}
\description{
Test that graph contains cycles
}
