% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/primitives.R
\name{E}
\alias{E}
\alias{edges}
\alias{E.datagraph}
\alias{edges.datagraph}
\title{List all edges in a graph}
\usage{
E(x, ...)

edges(x, ...)

\method{E}{datagraph}(x, sorted = FALSE)

\method{edges}{datagraph}(x, sorted = FALSE)
}
\arguments{
\item{x}{datagraph object}

\item{...}{args passed to the S3 methods}

\item{sorted}{logical; sort the result?}
}
\description{
List all edges in a graph
}
