% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{foreach}
\alias{foreach}
\title{Run a function for each member.}
\usage{
foreach(x, f, ...)
}
\arguments{
\item{x}{see the S3 methods}

\item{f}{function}

\item{...}{args passed to the S3 methods}
}
\description{
Run a function for each member.
}
