% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{is.datagraph}
\alias{is.datagraph}
\title{Function to check if an object is a datagraph}
\usage{
is.datagraph(x)
}
\arguments{
\item{x}{an R object}
}
\description{
Function to check if an object is a datagraph
}
