% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/neighborhood.R
\name{neighbors}
\alias{neighbors}
\alias{neighbors_in}
\alias{neighbors_out}
\alias{neighborhood}
\alias{neighbors_in.datagraph_vertex}
\alias{neighbors_in.datagraph}
\alias{neighbors_out.datagraph_vertex}
\alias{neighbors_out.datagraph}
\alias{neighbors.datagraph}
\alias{neighborhood.datagraph}
\title{Adjacent vertices}
\usage{
neighbors(x, ...)

neighbors_in(x, ...)

neighbors_out(x, ...)

neighborhood(x, ...)

\method{neighbors_in}{datagraph_vertex}(x, names = FALSE)

\method{neighbors_in}{datagraph}(x, vertices, names = FALSE, useNames = TRUE)

\method{neighbors_out}{datagraph_vertex}(x, names = FALSE)

\method{neighbors_out}{datagraph}(x, vertices, names = FALSE, useNames = TRUE)

\method{neighbors}{datagraph}(x, vertices, mode = "all", names = FALSE)

\method{neighborhood}{datagraph}(x, vertices, order = 1000, mode = "all", names = FALSE)
}
\arguments{
\item{x}{see individual S3 methods}

\item{...}{}

\item{names}{logical; when TRUE only names of the vertices are returned, when FALSE -> vertex objects are returned}

\item{useNames}{logical; when TRUE, a named list/vector is returned}

\item{mode}{character; "all", "in" our "out}

\item{order}{integer; how far from the node to look}
}
\value{







}
\description{
neighbors - adjacent vertices

neighbors_in - adjacent vertices following incoming edges

neighbors_out - adjacent vertices following outgoing edges
}
