% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{remove_edge}
\alias{remove_edge}
\alias{remove_edges}
\title{Remove edges between vertices}
\usage{
remove_edge(x, ...)

remove_edges(x, ...)
}
\arguments{
\item{x}{see the S3 methods}

\item{...}{args passed to the S3 methods}
}
\description{
Remove edges between vertices
}
