% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{remove_vertex}
\alias{remove_vertex}
\title{Remove vertices from a x}
\usage{
remove_vertex(x, vertex, ...)
}
\arguments{
\item{x}{see the S3 methods}

\item{vertex}{a vertex object}

\item{...}{args passed to the S3 methods}
}
\description{
Remove vertices from a x
}
