% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/primitives.R
\name{remove_vertices}
\alias{remove_vertices}
\alias{remove_vertex.datagraph}
\alias{remove_vertex.character}
\alias{remove_vertices.datagraph}
\title{Remove vertices from a graph}
\usage{
remove_vertices(x, vertices, ...)

\method{remove_vertex}{datagraph}(x, vertex, ...)

\method{remove_vertex}{character}(x, graph)

\method{remove_vertices}{datagraph}(x, vertices, ...)
}
\arguments{
\item{x}{see the S3 methods}

\item{vertices}{vertex ids}

\item{...}{passed to individual methods}

\item{vertex}{vertex id}

\item{graph}{datagraph object}
}
\description{
Remove vertices from a graph
}
