% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/union.R
\name{union.datagraph}
\alias{union.datagraph}
\alias{union_list_of_graphs}
\title{Union multiple datagraphs into a single datagraph object}
\usage{
\method{union}{datagraph}(...)

union_list_of_graphs(x)
}
\arguments{
\item{...}{args passed to union_list_of_graphs}

\item{x}{list of datagraph objects}
}
\description{
Union multiple datagraphs into a single datagraph object
}
