% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/primitives.R
\name{V}
\alias{V}
\alias{vertices}
\alias{V.datagraph}
\alias{vertices.datagraph}
\title{List all vertices in a graph}
\usage{
V(x, ...)

vertices(x, ...)

\method{V}{datagraph}(x, sorted = FALSE)

\method{vertices}{datagraph}(x, sorted = FALSE)
}
\arguments{
\item{x}{see the S3 methods}

\item{...}{args passed to the S3 methods}

\item{sorted}{logical; when TRYE the result is sorted.}
}
\description{
List all vertices in a graph
}
