% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{remove_edge}
\alias{remove_edge}
\alias{remove_edges}
\title{Remove edges between vertices}
\usage{
remove_edge(x, ...)

remove_edges(x, ...)
}
\arguments{
\item{...}{}
}
\value{



}
\description{
Remove edges between vertices
}
