% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{remove_vertices}
\alias{remove_vertices}
\title{Remove vertices from a graph}
\usage{
remove_vertices(x, vertices, ...)
}
\arguments{
\item{...}{passed to individual methods}
}
\value{

}
\description{
Remove vertices from a graph
}
