% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{remove_vertices}
\alias{remove_vertices}
\title{Remove vertices from a graph}
\usage{
remove_vertices(graph, vertices, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Remove vertices from a graph
}
