% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{remove_vertices}
\alias{remove_vertices}
\title{Remove vertices from a graph}
\usage{
remove_vertices(x, vertices, ...)
}
\arguments{
\item{x}{see the S3 methods}

\item{vertices}{a vector of vertices' ids}

\item{...}{passed to individual methods}
}
\description{
Remove vertices from a graph
}
